working on module-consumer friendly design
