--- data/schema.js
@@ -1,1 +1,0 @@
-export { StarWarsSchema as default } from './swapiSchema'
