+++ package.json
@@ -2,8 +2,14 @@
   "name": "starwars-server",
   "version": "0.1.2",
   "description": "GraphQL Server setup with an example Star Wars schema",
+  "main": "./lib/server.js",
+  "files": [
+    "data.json",
+    "schema.graphql"
+  ]
   "scripts": {
-    "start": "nodemon ./server.js --exec babel-node",
+    "start": "nodemon ./src/server.js --exec babel-node",
+    "build": "babel ./src/ --out-dir ./lib/",
     "lint": "eslint ."
   },
   "repository": {
