+++ src/data/schema.js
@@ -1,0 +1,2 @@
+import { StarWarsSchema } from "./swapiSchema";
+export default StarWarsSchema;
